+++ ro-ro-monitor/src/main/java/com/company/roro/entity/RouteOtdConfig.java
@@ -20,51 +20,51 @@ public class RouteOtdConfig {
     /** 线路ID，关联 route_dict 表 */
     private Integer routeId;
 
-    // ==================== 7段标准OTD时效（单位：小时） ====================
+    // ==================== 7段标准OTD时效（单位：小时，支持小数） ====================
 
     /** 未出库 → 集港在途 标准时效 */
-    private Integer notDepartedOtd;
+    private Double notDepartedOtd;
 
     /** 集港在途 → 已集港待装船 标准时效 */
-    private Integer toPortOtd;
+    private Double toPortOtd;
 
     /** 已集港待装船 → 水运在途 标准时效 */
-    private Integer atPortWaitOtd;
+    private Double atPortWaitOtd;
 
     /** 水运在途 → 已到港待卸船 标准时效 */
-    private Integer onSeaOtd;
+    private Double onSeaOtd;
 
     /** 已到港待卸船 → 已卸船待分拨 标准时效 */
-    private Integer atDestWaitOtd;
+    private Double atDestWaitOtd;
 
     /** 已卸船待分拨 → 分拨在途 标准时效 */
-    private Integer unloadWaitDispatchOtd;
+    private Double unloadWaitDispatchOtd;
 
     /** 分拨在途 → 已到达 标准时效 */
-    private Integer dispatchingOtd;
+    private Double dispatchingOtd;
 
-    // ==================== 7段预警时效（单位：小时） ====================
+    // ==================== 7段预警时效（单位：小时，支持小数） ====================
 
     /** 未出库预警时效 */
-    private Integer notDepartedWarn;
+    private Double notDepartedWarn;
 
     /** 集港在途预警时效 */
-    private Integer toPortWarn;
+    private Double toPortWarn;
 
     /** 已集港待装船预警时效 */
-    private Integer atPortWaitWarn;
+    private Double atPortWaitWarn;
 
     /** 水运在途预警时效 */
-    private Integer onSeaWarn;
+    private Double onSeaWarn;
 
     /** 已到港待卸船预警时效 */
-    private Integer atDestWaitWarn;
+    private Double atDestWaitWarn;
 
     /** 已卸船待分拨预警时效 */
-    private Integer unloadWaitDispatchWarn;
+    private Double unloadWaitDispatchWarn;
 
     /** 分拨在途预警时效 */
-    private Integer dispatchingWarn;
+    private Double dispatchingWarn;
 
     /** 是否启用：1启用，0禁用 */
     private Integer isActive;
